requirements.txt
groq
PyYAML
torch
fastapi
uvicorn
Flask
gTTS
openai
